fix warning: Since PHP 7.2.0 count() yields a warning on invalid countable types such as null

+++ app/controllers/AnimalsController.php
@@ -260,7 +260,7 @@ class AnimalsController extends BaseController
 
         /* Delete the selected pictures */
         $picstodelete = Input::get('picstodelete');
-        if (count($picstodelete) > 0) {
+        if ($picstodelete !== null && count($picstodelete) > 0) {
             foreach($picstodelete as $picId) {
                 $pic = AnimalPic::findOrFail($picId);
                 unlink($images_dir . $pic->filename); // "Delete" the file.
